requirements.txt
ansicon==1.89.0
APScheduler==3.7.0
arrow==0.17.0
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
blessed==1.17.12
bs4==0.0.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
DateTime==4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-apscheduler==0.5.2
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-picklefield==3.0.1
django-q==1.3.4
djangorestframework==3.12.2
gunicorn==20.0.4
heroku==0.1.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jinxed==1.0.1
jmespath==0.10.0
lxml==4.6.2
oauthlib==3.1.0
parsel==1.6.0
Protego==0.1.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==2.0.1
pyOpenSSL==20.0.0
python-dateutil==2.8.1
python-scrapyd-api==2.1.2
python3-openid==3.2.0
pytz==2020.4
queuelib==1.5.0
requests==2.25.1
requests-oauthlib==1.3.0
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
service-identity==18.1.0
six==1.15.0
social-auth-core==4.0.3
soupsieve==2.0.1
sqlparse==0.4.1
Twisted==20.3.0
tzlocal==2.1
urllib3==1.26.3
w3lib==1.22.0
wcwidth==0.2.5
whitenoise==5.2.0
zope.interface==5.2.0
